courseapp/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.39.0
cryptography==42.0.5
Django==5.0.3
django-ckeditor==6.7.1
django-filter==24.1
django-js-asset==2.2.0
django-oauth-toolkit==2.3.0
djangorestframework==3.15.0
drf-yasg==1.21.7
idna==3.6
inflection==0.5.1
jwcrypto==1.5.6
Markdown==3.6
oauthlib==3.2.2
packaging==24.0
pillow==10.2.0
pycparser==2.22
PyMySQL==1.1.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
